% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metacells.R
\name{calculate_metacells}
\alias{calculate_metacells}
\title{calculate_metacells}
\usage{
calculate_metacells(sce_query, annotation_col = NULL, gamma = 10, n.pc = 10)
}
\arguments{
\item{sce_query}{SingleCellExperiment object with logcounts to calculate metacells on}

\item{annotation_col}{name of annotation column in sce_query, set to NULL if no annotation should be transferred to metacell level}

\item{gamma}{set the graining level for the metacells, aka proportion of single cells to meta cells}

\item{n.pc}{number of PCAs to be used during the calculation of the metacell identities, default is 10}
}
\value{
SCE with logcounts on metacell level, also clusters, PCA & tSNE
}
\description{
calculate_metacells
}
\examples{
# create a SCE on metacell level without prior annotation
#load example SCE from iUSEiSEE package

sce_annotated <- readRDS(
 file = system.file("datasets", "sce_pbmc3k.RDS", package = "iUSEiSEE"))

sce_metacells <- calculate_metacells(sce_query = sce_annotated,
                                     annotation_col = NULL)

sce_metacells


}
