% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToSCN.R
\name{convertToSCN}
\alias{convertToSCN}
\title{convertToSCN}
\usage{
convertToSCN(sce_object, exp_type = "counts")
}
\arguments{
\item{sce_object}{SCE object to be converted to SCN format}

\item{exp_type}{type of assay data in the SCE (eg "counts", "logcounts", "normcounts" etc)}
}
\value{
list with sampTab (the metadata) and expDat (the expression matrix),
the names of the rows (aka cells) will be stored in the sample_name column
}
\description{
extract sampTab and expDat sce object into regular S3 objects
}
