% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scBaptism_majority.R
\name{get_winner}
\alias{get_winner}
\title{scBaptism_majority}
\usage{
get_winner(input_vector)
}
\arguments{
\item{input_vector}{vector to do the majority voting on}
}
\value{
sce_query: SCE object with an added metadata column for the majority annotation
}
\description{
scBaptism_majority
}
\examples{

#loading example dataset
library(iUSEiSEE)

sce_annotated <- readRDS(file = system.file("datasets", "sce_pbmc3k.RDS", package = "iUSEiSEE"))

}
