% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majority_voting.R
\name{majority_vote}
\alias{majority_vote}
\title{majority_vote}
\usage{
majority_vote(sce_query, anno_columns = NULL, tie_breaker)
}
\arguments{
\item{sce_query}{SingleCellExperiment object with multiple annotations}

\item{anno_columns}{Vector with names of annotation columns, if NULL all columns starting with "scb" will be used}

\item{tie_breaker}{choose what happens with a tie between the annotations, options are "na" and "first", default is "na"}
}
\value{
sce_query: SCE object with an added metadata column for the majority annotation
}
\description{
majority_vote
}
\examples{

# loading example dataset, will also serve as a reference
library(iUSEiSEE)
sce_annotated <- readRDS(file = system.file("datasets",
                                            "sce_pbmc3k.RDS",
                                            package = "iUSEiSEE"))

# use scBaptism to get several single cell annotations
sce_annotated <- run_scBaptism(sce_query = sce_annotated,
                               anno_methods = c("CelliDref", "SingleR", "scClassify"),
                               reference = sce_annotated,
                               ref_labs = "labels_main")

# perform a majority vote on all annotations obtained through scBaptism
# with choosing the first annotation in case of a tie
sce_annotated <- majority_vote(sce_query = sce_annotated,
                               anno_columns = NULL,
                               tie_breaker = "first")

# plot the majority vote annotation
scater::plotTSNE(sce_annotated, color_by = "scb_majority_labels")



}
