% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_allivial}
\alias{plot_allivial}
\title{plot_alluvial}
\usage{
plot_allivial(sce_query, first_tool, second_tool, threshold = 0.02)
}
\arguments{
\item{sce_query}{SingleCellExperiment object with the annotations to plot in the colData}

\item{first_tool}{name of the first annotation column in the colData of sce_query}

\item{second_tool}{name of the second annotation column in the colData of sce_query}

\item{threshold}{number defining below which percentage of occurrence cells get filtered out (default 0.02, so 2\%)}
}
\value{
alluvial plot comparing the annotation of two selected annotation columns
}
\description{
plot_alluvial
}
