% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{plot_confusion_matrix}
\usage{
plot_confusion_matrix(sce_query, first_tool, second_tool)
}
\arguments{
\item{sce_query}{SingleCellExperiment object with the annotations to plot in the colData}

\item{first_tool}{name of the first annotation column in the colData of sce_query}

\item{second_tool}{name of the second annotation column in the colData of sce_query}
}
\value{
confusion matrix comparing the annotation of two selected annotation columns
}
\description{
plot_confusion_matrix
}
