% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(sce_query, first_tool, second_tool)
}
\arguments{
\item{sce_query}{SingleCellExperiment object with the annotations to plot in the colData}

\item{first_tool}{name of the first annotation column in the colData of sce_query}

\item{second_tool}{name of the second annotation column in the colData of sce_query}
}
\value{
heatmap comparing the annotation of two selected annotation columns
}
\description{
plot_heatmap
}
\examples{
#load example SCE from iUSEiSEE package

sce_annotated <- readRDS(
 file = system.file("datasets", "sce_pbmc3k.RDS", package = "iUSEiSEE"))

#plot two of the exisiting annotations

plot <- plot_heatmap(sce_query = sce_annotated,
                              first_tool = "labels_main",
                              second_tool = "labels_fine")

plot

}
