% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_multiple_tSNE}
\alias{plot_multiple_tSNE}
\title{plot_multiple_tSNE}
\usage{
plot_multiple_tSNE(sce_query, labels_vector)
}
\arguments{
\item{sce_query}{SingleCellExperiment object with the annotations to plot in the colData}

\item{labels_vector}{Vector with names of the colData}
}
\value{
plot with all selected annotations plotted with tSNE dimensionality reduction
}
\description{
plot_multiple_tSNE
}
\examples{
#load example SCE from iUSEiSEE package

sce_annotated <- readRDS(
 file = system.file("datasets", "sce_pbmc3k.RDS", package = "iUSEiSEE"))

#plot two of the exisiting annotations

plot <- plot_multiple_tSNE(sce_query = sce_annotated,
                          labels_vector = c("labels_main", "labels_fine"))

plot


}
