% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_CelliDref.R
\name{run_CelliDref}
\alias{run_CelliDref}
\title{run_CelliDref}
\usage{
run_CelliDref(
  sce_query,
  reference,
  ref_labs,
  reduction = "MCA",
  n.features = 200,
  CelliD_features = NULL,
  CelliD_dims = seq(50),
  minSize = 10,
  log.trans = TRUE,
  p.adjust = TRUE,
  return_extra_info = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{sce_query}{SCE to be annotated}

\item{reference}{SCE with annotations}

\item{ref_labs}{existing annotation in the reference}

\item{reduction}{name of the MCA reduction}

\item{n.features}{integer of top n features to consider for hypergeometric test}

\item{CelliD_features}{vector of features to calculate the gene ranking by default will take everything in the selected mca reduction.}

\item{CelliD_dims}{MCA dimensions to use to compute n.features top genes.}

\item{minSize}{minimum number of overlapping genes in geneset and}

\item{log.trans}{if TRUE tranform the pvalue matrix with -log10 and convert it to sparse matrix}

\item{p.adjust}{if TRUE apply Benjamini Hochberg correction to p-value}

\item{return_extra_info}{if TRUE, adds additional metadata from the annotation}

\item{verbose}{display message after annotation is finished}
}
\value{
sce_query a SingleCellExperiment object, with the extra info on the
annotated cells
}
\description{
run_CelliDref
}
\examples{

library(iUSEiSEE)
library(dplyr)

# load SCE from iUSEiSEE

sce_annotated <- readRDS(file = system.file("datasets", "sce_pbmc3k.RDS", package = "iUSEiSEE"))

# run the annotation
sce_annotated <- run_CelliDref(sce_query = sce_annotated,
                               reference = sce_annotated,
                               ref_labs = "labels_main")

# plot the existing annotation with scater(t-SNE)
scater::plotTSNE(sce_annotated, color_by = "scb_CelliDref_labels")


}
\seealso{
Other hybrid family: 
\code{\link{run_CelliDmk}()}
}
\concept{hybrid family}
