% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_clustifyr.R
\name{run_clustifyr}
\alias{run_clustifyr}
\title{run_clustifyr}
\usage{
run_clustifyr(
  sce_query,
  reference,
  ref_labs,
  query_genes = NULL,
  per_cell = FALSE,
  n_perm = 0,
  compute_method = "spearman",
  pseudobulk_method = "mean",
  dr = "umap",
  obj_out = TRUE,
  seurat_out = obj_out,
  vec_out = FALSE,
  threshold = "auto",
  rm0 = FALSE,
  rename_prefix = NULL,
  exclude_genes = c(),
  metadata = NULL,
  organism = "hsapiens",
  plot_name = NULL,
  rds_name = NULL,
  expand_unassigned = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{sce_query}{SCE to be annotated}

\item{reference}{SCE object that acts as a reference}

\item{ref_labs}{List of gene labels or column from the references colData}

\item{per_cell}{if true run per cell, otherwise per cluster.}

\item{n_perm}{number of permutations, set to 0 by default}

\item{compute_method}{method(s) for computing similarity scores}

\item{pseudobulk_method}{method used for summarizing clusters, options are mean (default), median, truncate (10\% truncated mean), or trimean, max, min}

\item{dr}{stored dimension reduction}

\item{obj_out}{whether to output object instead of cor matrix}

\item{seurat_out}{output cor matrix or called seurat object (deprecated, use obj_out instead)}

\item{vec_out}{only output a result vector in the same order as metadata}

\item{threshold}{identity calling minimum correlation score threshold, only used when obj_out = TRUE}

\item{rm0}{consider 0 as missing data, recommended for per_cell}

\item{rename_prefix}{prefix to add to type and r column names}

\item{exclude_genes}{a vector of gene names to throw out of query}

\item{organism}{for GO term analysis, organism name: human - 'hsapiens', mouse - 'mmusculus'}

\item{plot_name}{name for saved pdf, if NULL then no file is written (default)}

\item{rds_name}{name for saved rds of rank_diff, if NULL then no file is written (default)}

\item{expand_unassigned}{test all ref clusters for unassigned results}

\item{verbose}{display message after annotation is finished}

\item{query_genesnA}{vector of genes of interest to compare. If NULL, then common genes between the expr_mat and ref_mat will be used for comparision.}

\item{n_genes}{number of genes limit for Seurat variable genes, by default 1000, set to 0 to use all variable genes (generally not recommended)}

\item{low_threshold_cell}{option to remove clusters with too few cells}

\item{if_log}{input data is natural log, averaging will be done on unlogged data}

\item{use_var_genes}{if providing a seurat object, use the variable genes (stored in seurat_object@var.genes) as the query_genes.}
}
\value{
sce_query a SingleCellExperiment object, with the extra info on the
annotated cells
}
\description{
run_clustifyr
}
\examples{

library(iUSEiSEE)
library(Seurat)
library(dplyr)

#load SCE from iUSEiSEE

sce_annotated <- readRDS(file = system.file("datasets", "sce_pbmc3k.RDS", package = "iUSEiSEE"))

#run the annotation
sce_annotated <- run_clustifyr(sce_query = sce_annotated,
                               reference = sce_annotated,
                               ref_labs = "labels_main")

#plot the existing annotation with scater(t-SNE)
scater::plotTSNE(sce_annotated, color_by = "scb_clustifyr_labels")


}
\seealso{
Other reference-based family: 
\code{\link{run_Seurat}()},
\code{\link{run_SingleR}()}
}
\concept{reference-based family}
