% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rSuperCT.R
\name{run_rSuperCT}
\alias{run_rSuperCT}
\title{run_rSuperCT}
\usage{
run_rSuperCT(
  sce_query,
  SuperCT_classifier = "generic_38celltypes",
  species = "human",
  return_extra_info = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{sce_query}{SCE to be annotated}

\item{SuperCT_classifier}{one of the three classifiers provided on SuperCT repo}

\item{species}{"human" or "mouse"}

\item{verbose}{display message after annotation is finished}
}
\value{
sce_query a SingleCellExperiment object, with the extra info on the
annotated cells
}
\description{
run_rSuperCT
}
\examples{


library(iUSEiSEE)
library(Seurat)
library(dplyr)


# load SCE from iuseisee

sce_annotated <- readRDS(
 file = system.file("datasets", "sce_pbmc3k.RDS", package = "iUSEiSEE"))



#run the annotation
#sce_annotated <- run_rSuperCT(sce_annotated, markers_lists)

# plot the existing annotation with scater(t-SNE)
scater::plotTSNE(sce_annotated, color_by = "labels_main")


}
\concept{deep learning family

function is commented out since it does not really fit with our vision of the package
(mainly is does not allow to use ones own reference)
but the function work fine and can be used if desired}
