% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scBaptism.R
\name{run_scBaptism}
\alias{run_scBaptism}
\title{run_scBaptism}
\usage{
run_scBaptism(
  sce_query = NULL,
  anno_methods = NULL,
  markers_list = NULL,
  reference = NULL,
  ref_labs = NULL,
  verbose = FALSE,
  return_extra_info = FALSE,
  ...
)
}
\arguments{
\item{sce_query}{SCE to be annotated}

\item{anno_methods}{vector with all annotation methods to be used for annotation, default is all methods}

\item{markers_list}{list of marker genes}

\item{reference}{SingleCellExperiment object with annotation information column}

\item{ref_labs}{name of column with annotation information}

\item{verbose}{display message after each annotation is finished}

\item{return_extra_info}{if TRUE, adds additional metadata from the annotation when available}

\item{...}{further arguments passed to other methods}
}
\value{
sce_query : a SingleCellExperiment object, with the extra info on the
annotated cells
}
\description{
run_scBaptism
}
