% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scClassify.R
\name{run_scClassify}
\alias{run_scClassify}
\title{run_scClassify}
\usage{
run_scClassify(
  sce_query,
  reference,
  ref_labs,
  return_extra_info = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{sce_query}{SCE to be annotated}

\item{reference}{SCE object that acts as a reference}

\item{return_extra_info}{if TRUE, adds additional metadata from the annotation}

\item{verbose}{display message after annotation is finished}

\item{ref_labels}{Column from the references colData}
}
\value{
sce_query a SingleCellExperiment object, with the extra info on the
annotated cells
}
\description{
run_scClassify
}
\examples{

library(iUSEiSEE)
library(dplyr)
# load SCE from iUSEiSEE

sce_annotated <- readRDS(file = system.file("datasets", "sce_pbmc3k.RDS", package = "iUSEiSEE"))

# run the annotation
sce_annotated <- run_scClassify(sce_query = sce_annotated,
                                reference = sce_annotated,
                                ref_labs = "labels_main")

# plot the existing annotation with scater(t-SNE)
scater::plotTSNE(sce_annotated, color_by = "scb_scClassify_labels")


}
\concept{reference family}
