% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_anno_SCINA.R
\name{run_SCINA}
\alias{run_SCINA}
\title{Title}
\usage{
run_SCINA(query, reference, alina, ...)
}
\arguments{
\item{alina}{}
}
\value{
A SingleCellExperiment object, with the extra info on the
annotated cells
}
\description{
Title
}
\examples{

# TODO
}
