% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scmap.R
\name{run_scmap}
\alias{run_scmap}
\title{run_scmap}
\usage{
run_scmap(
  sce_query,
  reference,
  ref_labs,
  threshold_Cluster = 0.7,
  threshold_Cell2Cluster = 0.5,
  w_Cell = 10,
  w_Cell2Cluster = 3,
  return_extra_info = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{sce_query}{SCE to be annotated}

\item{reference}{SCE object that acts as a reference}

\item{ref_labs}{Column from the references colData}

\item{threshold_Cluster}{threshold on similarity for cluster-wise annotation}

\item{threshold_Cell2Cluster}{the threshold which the maximum similarity between the query and a reference cell must exceed for the cell-type to be assigned}

\item{w_Cell}{a positive integer specifying the number of nearest neighbors to find during cell indexing}

\item{w_Cell2Cluster}{an integer specifying the number of nearest neighbors to find during sc annotation}

\item{return_extra_info}{if TRUE, adds additional metadata from the annotation}

\item{verbose}{display message after annotation is finished}
}
\value{
sce_query a SingleCellExperiment object, with the extra info on the
annotated cells
}
\description{
run_scmap
}
\examples{

library(iUSEiSEE)
library(Seurat)
library(dplyr)

#load SCE from iUSEiSEE

sce_annotated <- readRDS(file = system.file("datasets", "sce_pbmc3k.RDS", package = "iUSEiSEE"))

#run the annotation
sce_annotated <- run_scmap(sce_query= sce_annotated,
                          reference = sce_annotated,
                           ref_labs = "labels_main")

#plot the new annotations with scater(t-SNE)
#cell wise annotation
scater::plotTSNE(sce_annotated, color_by = "scb_scmap_labels")

#cluster wise annotation
scater::plotTSNE(sce_annotated, color_by = "scb_scmap_clusterlabels")

}
\seealso{
Other classical machine learning family: 
\code{\link{run_scPred}()}
}
\concept{classical machine learning family}
