% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scPred.R
\name{run_scPred}
\alias{run_scPred}
\title{run_scPred}
\usage{
run_scPred(
  sce_query,
  reference,
  ref_lab,
  threshold = 0.55,
  max.iter.harmony = 20,
  recompute_alignment = TRUE,
  seed = 66,
  return_extra_info = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{sce_query}{SCE to be annotated}

\item{reference}{SCE object that acts as a reference}

\item{return_extra_info}{if TRUE, adds additional metadata from the annotation}

\item{verbose}{display message after annotation is finished}

\item{ref_labels}{Column from the references colData}
}
\value{
sce_query a SingleCellExperiment object, with the extra info on the
annotated cells

export
}
\description{
run_scPred
}
\examples{


library(iUSEiSEE)
library(dplyr)

# load SCE from iUSEiSEE

sce_annotated <- readRDS(file = system.file("datasets", "sce_pbmc3k.RDS", package = "iUSEiSEE"))

#run the annotation
#sce_annotated <- run_scPred()

# plot the existing annotation with scater(t-SNE)
scater::plotTSNE(sce_annotated, color_by = "labels_main")


}
\seealso{
Other classical machine learning family: 
\code{\link{run_SingleCellNet}()},
\code{\link{run_scmap}()}
}
\concept{classical machine learning family}
