% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SingleCellNet.R
\name{run_SingleCellNet}
\alias{run_SingleCellNet}
\title{run_SingleCellNet}
\usage{
run_SingleCellNet(
  sce_query,
  exp_type_query,
  reference,
  exp_type_ref,
  ref_lab,
  return_extra_info = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{sce_query}{SCE to be annotated}

\item{reference}{SCE object that acts as a reference}

\item{return_extra_info}{if TRUE, adds additional metadata from the annotation}

\item{verbose}{display message after annotation is finished}

\item{ref_labels}{Column from the references colData}
}
\value{
sce_query a SingleCellExperiment object, with the extra info on the
annotated cells
}
\description{
run_SingleCellNet
}
\examples{


library(iUSEiSEE)
library(dplyr)

# load SCE from iUSEiSEE

sce_annotated <- readRDS(file = system.file("datasets", "sce_pbmc3k.RDS", package = "iUSEiSEE"))

#run the annotation
#sce_annotated <- run_SingleCellNet(sce_annotated, markers_lists)

# plot the existing annotation with scater(t-SNE)
scater::plotTSNE(sce_annotated, color_by = "labels_main")


}
\seealso{
Other classical machine learning family: 
\code{\link{run_scPred}()},
\code{\link{run_scmap}()}
}
\concept{classical machine learning family}
