% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SingleR.R
\name{run_SingleR}
\alias{run_SingleR}
\title{run_xx}
\usage{
run_SingleR(sce_query, reference, ref_labels, extra_data = TRUE, ...)
}
\arguments{
\item{sce_query}{SCE object to be annotated}

\item{reference}{SCE object that acts as a reference}

\item{ref_labels}{List of gene labels or column from the refernces colData}
}
\value{
sce_query : a SingleCellExperiment object, with the extra info on the
annotated cells
}
\description{
run_xx
}
\examples{

#############exsample not yet complete
library(iUSEiSEE)
library(Seurat)
library(dplyr)


# load SCE from iuseisee

sce_annotated <- readRDS(
 file = system.file("datasets", "sce_pbmc3k.RDS", package = "iUSEiSEE")
)

# plot the existing annotation with scater(t-SNE)
scater::plotTSNE(sce_annotated, color_by = "labels_main")


}
\concept{reference-based family}
