% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metacells.R
\name{calculate_metacells}
\alias{calculate_metacells}
\title{calculate_metacells}
\usage{
calculate_metacells(sce_query, gamma = 10)
}
\arguments{
\item{sce_query}{SingleCellExperiment object with logcounts to calculate metacells on}

\item{gamma}{set the graining level for the metacells, aka proportion of single cells to meta cells}
}
\value{
SCE with logcounts on metacell level, also clusters, PCA & tSNE
}
\description{
calculate_metacells
}
\examples{
#load example SCE from iUSEiSEE package

sce_annotated <- readRDS(
 file = system.file("datasets", "sce_pbmc3k.RDS", package = "iUSEiSEE"))

sce_metacells <- calculate_metacells(sce_query = sce_annotated,
                                     gamma = 10)

sce_metacells


}
