% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_CelliD.R, R/run_CelliDmk.R
\name{run_CelliDmk}
\alias{run_CelliDmk}
\title{run_CelliDmk}
\usage{
run_CelliDmk(
  sce_query,
  markers_list,
  return_extra_info = FALSE,
  verbose = FALSE,
  ...
)

run_CelliDmk(
  sce_query,
  markers_list,
  return_extra_info = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{sce_query}{SCE to be annotated}

\item{markers_list}{List of marker genes}

\item{return_extra_info}{if TRUE, adds additional metadata from the annotation}

\item{verbose}{display message after annotation is finished}
}
\value{
sce_query a SingleCellExperiment object, with the extra info on the
annotated cells

export

sce_query a SingleCellExperiment object, with the extra info on the
annotated cells

export
}
\description{
run_CelliDmk

run_CelliDmk
}
\examples{


library(iUSEiSEE)
library(dplyr)

# load SCE from iUSEiSEE

sce_annotated <- readRDS(file = system.file("datasets", "sce_pbmc3k.RDS", package = "iUSEiSEE"))

#run the annotation
#sce_annotated <- run_tool(sce_annotated, markers_lists)

# plot the existing annotation with scater(t-SNE)
scater::plotTSNE(sce_annotated, color_by = "scb")




library(iUSEiSEE)
library(dplyr)

load SCE from iUSEiSEE

sce_annotated <- readRDS(file = system.file("datasets", "sce_pbmc3k.RDS", package = "iUSEiSEE"))

run the annotation
sce_annotated <- run_tool(sce_annotated, markers_lists)

plot the existing annotation with scater(t-SNE)
scater::plotTSNE(sce_annotated, color_by = "scb_CelliDmk_labels")


}
\seealso{
Other hybrid family: 
\code{\link{run_CelliDref}()}

Other hybrid family: 
\code{\link{run_CelliDref}()}
}
\concept{hybrid family}
