% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_CIA.R
\name{run_CIA}
\alias{run_CIA}
\title{run_CIA}
\usage{
run_CIA(
  sce_query,
  markers_list,
  similarity_threshold = 0,
  column_name = "scb_CIA_res",
  n_cpus = 1
)
}
\arguments{
\item{sce_query}{SCE to be annotated}

\item{markers_list}{List of marker genes}
}
\value{
sce_query : a SingleCellExperiment object, with the extra info on the
annotated cells
}
\description{
run_CIA
}
\seealso{
Other marker family: 
\code{\link{run_SCINA}()}
}
\concept{marker family}
