% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_clustifyr.R
\name{run_clustifyr}
\alias{run_clustifyr}
\title{run_clustifyr}
\usage{
run_clustifyr(sce_query, reference, ref_labs, verbose = FALSE, ...)
}
\arguments{
\item{sce_query}{SCE to be annotated}

\item{reference}{SCE object that acts as a reference}

\item{ref_labs}{List of gene labels or column from the references colData}

\item{verbose}{display message after annotation is finished}
}
\value{
sce_query a SingleCellExperiment object, with the extra info on the
annotated cells

export
}
\description{
run_clustifyr
}
\examples{

library(iUSEiSEE)
library(Seurat)
library(dplyr)

#load SCE from iUSEiSEE

sce_annotated <- readRDS(file = system.file("datasets", "sce_pbmc3k.RDS", package = "iUSEiSEE"))

#run the annotation
sce_annotated <- run_clustifyr(sce_query = sce_annotated,
                               reference = sce_annotated,
                               ref_labs = "labels_main")

#plot the existing annotation with scater(t-SNE)
scater::plotTSNE(sce_annotated, color_by = "scb_clustifyr_labels")


}
\seealso{
Other reference-based family: 
\code{\link{run_Seurat}()},
\code{\link{run_SingleR}()}
}
\concept{reference-based family}
