% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scBaptism.R
\name{run_scBaptism}
\alias{run_scBaptism}
\title{run_scBaptism}
\usage{
run_scBaptism(
  sce_query = NULL,
  anno_methods = NULL,
  markers_list = NULL,
  reference = NULL,
  ref_labs = NULL,
  clusters = NULL,
  verbose = FALSE,
  return_extra_info = FALSE,
  ...
)
}
\arguments{
\item{sce_query}{SCE to be annotated}

\item{anno_methods}{vector with all annotation methods to be used for annotation, default is all methods}

\item{markers_list}{list of marker genes}

\item{reference}{SingleCellExperiment object with annotation information column}

\item{ref_labs}{name of column with annotation information}

\item{clusters}{name of clusters in the query}

\item{verbose}{display message after each annotation is finished}

\item{return_extra_info}{if TRUE, adds additional metadata from the annotation when available}

\item{...}{further arguments passed to other methods}
}
\value{
sce_query : a SingleCellExperiment object, with the extra info on the
annotated cells
}
\description{
run_scBaptism
}
\examples{
library(iUSEiSEE)
library(Seurat)
library(dplyr)


#load example SCE from iUSEiSEE package (also used as reference)

sce_annotated <- readRDS(
 file = system.file("datasets", "sce_pbmc3k.RDS", package = "iUSEiSEE"))

#get the markers list using Seurat
myseu <- Seurat::as.Seurat(sce_annotated)
myseu <- Seurat::ScaleData(myseu)
Seurat::Idents(myseu) <- "labels_main"

seu_all_markers <- Seurat::FindAllMarkers(myseu, test.use = "wilcox", only.pos = TRUE,
                                         min.pct = 0.25, logfc.threshold = 0.25)

top_k_markers <- 50

markers_lists <- seu_all_markers \%>\%
 dplyr::group_by(cluster) \%>\%
 dplyr::top_n(n = top_k_markers, wt = avg_log2FC)
markers_lists <- split(markers_lists$gene, markers_lists$cluster)

#run scBaptism to annotate with SingleR and CIA
anno_result <- run_scBaptism(sce_query = sce_annotated,
                             anno_methods = c("SCINA", "CIA"),
                             markers_list = markers_lists,
                             reference = sce_annotated,
                             ref_labs = "labels_main")

}
