% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SCINA.R
\name{run_SCINA}
\alias{run_SCINA}
\title{run_SCINA}
\usage{
run_SCINA(
  sce_query,
  markers_list,
  return_extra_info = FALSE,
  verbose = FALSE,
  max_iter = 1000,
  convergence_n = 12,
  convergence_rate = 0.999,
  sensitivity_cutoff = 0.9,
  rm_overlap = 1,
  allow_unknown = 1,
  log_file = "SCINA.log"
)
}
\arguments{
\item{sce_query}{SCE to be annotated}

\item{markers_list}{List of marker genes}

\item{return_extra_info}{if TRUE, adds additional metadata from the annotation}

\item{verbose}{display message after annotation is finished}

\item{max_iter}{max iterations allowed for EM algorithm}

\item{convergence_n}{stop the SCINA algorithm if during the last n rounds of iterations, cell type assignment keeps steady above the convergence_rate}

\item{convergence_rate}{percentage of cells for which the type assignment was stable for the last n rounds}

\item{sensitivity_cutoff}{cutoff to remove signatures whose cells types are deemed as non-existent at all in the data by the SCINA algorithm}

\item{rm_overlap}{binary deciding whether shared symbols in the markers_list will be deleted}

\item{allow_unknown}{if 1 unknown annotation is allowed, if 0 not allowed}

\item{log_file}{name of log file, that stores infor about SCINA algorithm}
}
\value{
sce_query: a SingleCellExperiment object, with the extra info on the
annotated cells
}
\description{
run_SCINA
}
\examples{
library(iUSEiSEE)
library(Seurat)
library(dplyr)


#load example SCE from iUSEiSEE package

sce_annotated <- readRDS(
 file = system.file("datasets", "sce_pbmc3k.RDS", package = "iUSEiSEE"))

#get the markers list using Seurat
myseu <- Seurat::as.Seurat(sce_annotated)
myseu <- Seurat::ScaleData(myseu)
Seurat::Idents(myseu) <- "labels_main"

seu_all_markers <- Seurat::FindAllMarkers(myseu, test.use = "wilcox", only.pos = TRUE,
                                         min.pct = 0.25, logfc.threshold = 0.25)

top_k_markers <- 50

markers_lists <- seu_all_markers \%>\%
 dplyr::group_by(cluster) \%>\%
 dplyr::top_n(n = top_k_markers, wt = avg_log2FC)
markers_lists <- split(markers_lists$gene, markers_lists$cluster)

# run the annotation
sce_annotated <- run_SCINA(sce_query = sce_annotated, markers_list =  markers_lists)

# plot the existing annotation with scater(t-SNE)
scater::plotTSNE(sce_annotated, color_by = "scb_SCINA_labels")

}
\seealso{
Other marker family: 
\code{\link{run_CIA}()}
}
\concept{marker family}
