% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SCINA.R
\name{run_SCINA}
\alias{run_SCINA}
\title{run_SCINA}
\usage{
run_SCINA(
  sce_query,
  markers_list,
  max_iter = 1000,
  convergence_n = 12,
  convergence_rate = 0.999,
  sensitivity_cutoff = 0.9,
  ...
)
}
\arguments{
\item{sce_query}{SCE to be annotated}

\item{markers_list}{List of marker genes}
}
\value{
description sce_query : a SingleCellExperiment object, with the extra info on the
annotated cells
}
\description{
run_SCINA
}
\examples{
# load libraries

library(iUSEiSEE)
library(Seurat)
library(dplyr)


# load example SCE from iuseisee package

sce_annotated <- readRDS(
 file = system.file("datasets", "sce_pbmc3k.RDS", package = "iUSEiSEE"))

# get the markers list
myseu <- Seurat::as.Seurat(sce_annotated)
myseu <- Seurat::ScaleData(myseu)
Seurat::Idents(myseu) <- "labels_main"

seu_all_markers <- Seurat::FindAllMarkers(myseu, test.use = "wilcox", only.pos = TRUE,
                                         min.pct = 0.25, logfc.threshold = 0.25)

top_k_markers <- 50

markers_lists <- seu_all_markers \%>\%
 group_by(cluster) \%>\%
 dplyr::top_n(n = top_k_markers, wt = avg_log2FC)
markers_lists <- split(markers_lists$gene, markers_lists$cluster)

# run the annotation
run_SCINA(sce_query = sce_annotated, markers_list =  markers_lists)

}
\seealso{
Other marker family: 
\code{\link{run_CIA}()}
}
\concept{marker family}
