% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scmap.R
\name{run_scmap}
\alias{run_scmap}
\title{run_scmap}
\usage{
run_scmap(
  sce_query,
  reference,
  ref_labels,
  return_extra_info = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{sce_query}{SCE to be annotated}

\item{reference}{SCE object that acts as a reference}

\item{ref_labels}{Column from the references colData}

\item{return_extra_info}{if TRUE, adds additional metadata from the annotation}

\item{verbose}{display message after annotation is finished}
}
\value{
sce_query a SingleCellExperiment object, with the extra info on the
annotated cells

export
}
\description{
run_scmap
}
\examples{


library(iUSEiSEE)
library(Seurat)
library(dplyr)


# load SCE from iuseisee

sce_annotated <- readRDS(file = system.file("datasets", "sce_pbmc3k.RDS", package = "iUSEiSEE"))



#run the annotation
#sce_annotated <- run_scmap(sce_annotated, markers_lists)

# plot the existing annotation with scater(t-SNE)
scater::plotTSNE(sce_annotated, color_by = "labels_main")


}
\seealso{
Other classical machine learning family: 
\code{\link{run_SingleCellNet}()},
\code{\link{run_scPred}()}
}
\concept{classical machine learning family}
