% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_Seurat.R
\name{run_seurat}
\alias{run_seurat}
\title{run_Seurat}
\usage{
run_seurat(sce_query)
}
\arguments{
\item{sce_query}{SCE to be annotated
param}
}
\value{
sce_query : a SingleCellExperiment object, with the extra info on the
annotated cells

export
}
\description{
run_Seurat
}
