% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SingleR.R
\name{run_SingleR}
\alias{run_SingleR}
\title{run_SinlgeR}
\usage{
run_SingleR(
  sce_query,
  reference,
  ref_labels,
  return_extra_info = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{sce_query}{SCE object to be annotated}

\item{reference}{SCE object that acts as a reference}

\item{ref_labels}{List of gene labels or column from the references colData}

\item{return_extra_info}{if TRuE, adds additional metadata from the annotation (delta.next, scores, prunded.labels)}

\item{verbose}{display message after annotation is finished}
}
\value{
sce_query : a SingleCellExperiment object, with the extra info on the
annotated cells
}
\description{
run_SinlgeR
}
\examples{

library(iUSEiSEE)
library(dplyr)


# load SCE from iuseisee

sce_annotated <- readRDS(
  file = system.file("datasets", "sce_pbmc3k.RDS", package = "iUSEiSEE")
)

#using the labels_main of the example to run another annotation with SingleR
sce_query <- run_SingleR(sce_query = sce_annotated, reference = sce_annotated, ref_labels = sce_annotated$labels_main)

# plot the existing annotation with scater(t-SNE)
scater::plotTSNE(sce_annotated, color_by = "scb_SingleR_res")


}
\seealso{
Other reference-based family: 
\code{\link{run_Seurat}()},
\code{\link{run_clustifyr}()}
}
\concept{reference-based family}
